Fix stuck spinner bug
In Sarcoidosis, when canceling out of an activity the spinner would stay up in front of the Activities tab, locking you out of the UI and requiring you to quit the app to be able to use it again. (At least, this was happening in the simulator.) It appears to have been due to attempting to dismiss the spinner while it was still in the process of animating in, so the solution was to start the process the spinner is meant to cover in the completion handler of that presentation animation.

If that makes any sense. If not, just look at the code change.